use rev returned by List in Watch
This commit solves the issue with Watch replaying all event starting from revision 1 if
the List call returns no values but there were some values under specified key in the
past revisions. It's the most reliable way to get all events from etcd without loosing
any of them and in the same time without receiving events that are not relevant.