etcd: Move out initial Get/List operation from goroutine in Watch methods
This commit moves out the Get and List operations
that are happening initially from the Watch goroutine
in order to avoid races if using the first current
revision we return from the channel. This also
improves error management as we used to just exit
the goroutine when the Get/List operation failed.

Signed-off-by: Alexandre Beslic <abeslic@abronan.com>