Dynamodb return all items
Due to:

If the total number of scanned items exceeds the maximum data set size limit of 1 MB, the scan stops and results are returned to the user as a LastEvaluatedKey value to continue the scan in a subsequent operation. The results also include the number of items exceeding the limit. A scan can result in no table data meeting the filter criteria.

Items should be retrieved from pages and decoded.

Use last zookeeper

Signed-off-by: Victor Castell <victor@victorcastell.com>